backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
werkzeug==2.2.3
PyPDF2==3.0.1
python-docx==0.8.11
# openai==0.27.0
anthropic==0.22.0
requests==2.28.2
beautifulsoup4==4.11.2
python-dotenv==1.0.0